% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad-outlier.R
\name{is_mad_outlier}
\alias{is_mad_outlier}
\title{Flag MAD outliers}
\usage{
is_mad_outlier(
  x,
  nmads = 3,
  direction = c("both", "lower", "upper"),
  constant = 1.4826,
  na_rm = TRUE,
  zero_mad = c("zero", "na", "error")
)
}
\arguments{
\item{x}{A numeric vector without infinite values.}

\item{nmads}{Number of MADs from the median.}

\item{direction}{Tail(s) for which to calculate finite limits.}

\item{constant}{Positive MAD consistency constant.}

\item{na_rm}{Remove missing values before calculation?}

\item{zero_mad}{Behavior when MAD is zero: neutral limits, missing limits,
or an error.}
}
\value{
A logical vector with the length and names of \code{x}.
}
\description{
Flag MAD outliers
}
\examples{
is_mad_outlier(c(1, 2, 2, 3, 100), direction = "upper")
}
