% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veryMAD-package.R
\docType{package}
\name{veryMAD-package}
\alias{veryMAD}
\alias{veryMAD-package}
\title{veryMAD: explicit MAD quality control and robust scaling}
\description{
A small, modality-agnostic toolkit for explicit observation-level QC flags
and robust MAD scaling. veryMAD never filters observations or infers causes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thokas99/veryMAD}
  \item \url{https://thokas99.github.io/veryMAD/}
  \item Report bugs at \url{https://github.com/Thokas99/veryMAD/issues}
}

}
\author{
\strong{Maintainer}: Thomas Sirchi \email{thomasirchi@gmail.com}

Authors:
\itemize{
  \item Thomas Sirchi \email{thomasirchi@gmail.com}
}

}
\keyword{internal}
