% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veryMAD-package.R
\docType{package}
\name{veryMAD-package}
\alias{veryMAD}
\alias{veryMAD-package}
\title{veryMAD: lightweight MAD methods for omics QC}
\description{
A lightweight R toolkit for robust MAD scores, dense expression-matrix
scaling and auditable quality-control flagging across bulk, single-cell and
other observation-level omics metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thokas99.github.io/veryMAD/}
}

}
\author{
\strong{Maintainer}: Thomas Sirchi \email{thomasirchi@gmail.com}

Authors:
\itemize{
  \item Thomas Sirchi \email{thomasirchi@gmail.com}
}

}
\keyword{internal}
