% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veryMAD-package.R
\docType{package}
\name{veryMAD-package}
\alias{veryMAD}
\alias{veryMAD-package}
\title{veryMAD: lightweight MAD methods for omics QC}
\description{
Pipe-friendly vector, matrix, data-frame, and explicit SeuratObject helpers
built around one consistent median absolute deviation policy.
}
\author{
\strong{Maintainer}: Thomas Sirchi \email{thomasirchi@gmail.com}

Authors:
\itemize{
  \item Thomas Sirchi \email{thomasirchi@gmail.com}
}

}
\keyword{internal}
